third_party/blink/web_tests/external/wpt/web-animations/timing-model/animations/seamlessly-updating-the-playback-rate-of-an-animation-expected.txt

This is a testharness.js-based test.
[FAIL] Updating the negative playback rate with the unresolved current time and a positive infinite associated effect end should not throw an exception
  assert_equals: We apply the pending playback rate immediately if the current time is unresolved expected -1 but got 1
Harness: the test ran to completion.
